+++ mythic_container/MythicGoRPC/__init__.py
@@ -1,4 +1,4 @@
-from .callback_display_to_real_id import *
+from .send_mythic_rpc_callback_display_to_real_id import *
 from .send_mythic_rpc_agentstorage_create import *
 from .send_mythic_rpc_agentstorage_remove import *
 from .send_mythic_rpc_agentstorage_search import *
@@ -55,7 +55,7 @@ from .send_mythic_rpc_task_search import *
 from .send_mythic_rpc_task_update import *
 from .send_mythic_rpc_token_create import *
 from .send_mythic_rpc_token_remove import *
-from .task_display_to_real_id import *
+from .send_mythic_rpc_task_display_to_real_id import *
 from .send_mythic_rpc_tag_create import *
 from .send_mythic_rpc_tag_search import *
 from .send_mythic_rpc_tagtype_get_or_create import *